% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\name{ci-object}
\alias{ci-object}
\title{A CI itnb-object}
\description{
A ci-object is a list of class \code{ci} containing the following:
\describe{
    \item{\code{ci}}{A list of elements corresponding to the mean, overdispersion, and inflation proportion. If a confidence level is specified it contains the quantiles for each parameter, otherwise all estimated parameters are returned.}
    \item{\code{parametric}}{Boolean indicating whether the bootstrap was parametric or non-parametric.}
    \item{\code{level}}{The specified confidence level; if a confidence level is not supplied it is \code{NA}.}
    \item{\code{nr_simulations}}{The number of bootstrap samples used to create the simulated confidence envelopes.}
}
}
