% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-optimise.R
\name{coef.itnb}
\alias{coef.itnb}
\title{Extract model coefficients}
\usage{
\method{coef}{itnb}(object, ...)
}
\arguments{
\item{object}{\link{itnb-object}.}

\item{...}{Additional arguments (see details).}
}
\value{
If \code{par} is left as \code{NULL} a list of all parameters will be returned, otherwise the function returns the parameter specified by \code{par}.
}
\description{
A function which extracts model coefficients from objects of class \code{itnb}.
}
\details{
The only additional argument used by the function is \code{par} used to specify which parameter should be returned by the function, i.e. it takes the values \code{"beta"}, \code{"alpha"}, or \code{"p"}.
}
