% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-rdp.R
\name{ditnb}
\alias{ditnb}
\title{ditnb}
\usage{
ditnb(
  x,
  mu,
  alpha,
  p = NULL,
  i = NULL,
  t = NULL,
  lower_tail = TRUE,
  return_log = FALSE
)
}
\arguments{
\item{x}{Numeric (>= 0): A vector of quantiles.}

\item{mu}{Numeric (> 0): The expected value.}

\item{alpha}{Numeric (>= 0): The overdispersion.}

\item{p}{Numeric (>= 0): The inflation proportion.}

\item{i}{Numeric (>= 0): The inflation point.}

\item{t}{Numeric (>= 0): The truncation point.}

\item{lower_tail}{TRUE/FALSE: should \eqn{P[X \leq x]} be returned in favour of\eqn{P[X \geq x]}?}

\item{return_log}{TRUE/FALSE: should the logarithm of the probabilities be returned?}
}
\value{
A vector the size as \code{x} containing the probability of each value.
}
\description{
The probability of each element in a vector \code{x} following the i-inflated t-truncated negative binomial distribution with parameters \code{mu}, \code{alpha}, and \code{p} which is inflated at \code{i} and truncated at \code{t}.
}
