% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb.R
\name{dkitnb}
\alias{dkitnb}
\title{dkitnb}
\usage{
dkitnb(x, k, t = NULL, mu, theta, p, return.log = FALSE)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{k}{The inflation point.}

\item{t}{The truncation point.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}
}
\value{
Numeric.
}
\description{
The density of a vector x following the k-inflated, t-t, negative binomial distribution.
}
