% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{ggplot.itnb}
\alias{ggplot.itnb}
\title{Plot parameter trace of itnb-object}
\usage{
\method{ggplot}{itnb}(itnb_object)
}
\arguments{
\item{itnb_object}{An object of the class 'itnb'.}
}
\value{
A ggplot.
}
\description{
A wrap function for ggplot2, plotting the Expectation-Maximisation parameter trace, which can returned by the \link{itnb_optimisation} function.
}
