% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktnb.R
\name{ggplot.ktnb}
\alias{ggplot.ktnb}
\title{Plot parameter trace of ktnb-object}
\usage{
\method{ggplot}{ktnb}(ktnb_object)
}
\arguments{
\item{ktnb_object}{An object of the class 'ktnb'.}
}
\value{
A ggplot.
}
\description{
A wrap function for ggplot2, plotting the Expectation-Maximisation parameter trace, which can returned by the \link{ktnb_optimisation} function.
}
