% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-bootstrap.R
\name{hist.itnb.ci}
\alias{hist.itnb.ci}
\title{Plot histograms of bootstrapped \link{itnb-object}}
\usage{
\method{hist}{itnb.ci}(x, which = NULL, ...)
}
\arguments{
\item{x}{\link{itnb.ci-object}.}

\item{which}{String: Indicating which parameter(s) to show. If left \code{NULL}, the function shows histograms of all parameters}

\item{...}{Additional arguments passed to the \link[graphics]{hist} function.}
}
\description{
A function plotting bootstrapped parameter estimates returned from the \link{confint.itnb} function.
}
