% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\name{itnb.ci-object}
\alias{itnb.ci-object}
\title{A summary.itnb-object}
\description{
A summary.itnb-object is a list of class \code{summary.itnb} containing the following:
\describe{
    \item{\code{formula}}{The \link{formula} supplied to the model.}
    \item{\code{link}}{The specified link function.}
    \item{\code{residuals}}{The residuals of the fitted model.}
    \item{\code{level}}{The specified confidence level.}
    \item{\code{coefficients}}{The estimated coefficients and bootstrap confidence intervals.}
    \item{\code{overdispersion}}{The estimated overdispersion and bootstrap confidence intervals.}
    \item{\code{inflation}}{The estimated inflation proportion and bootstrap confidence intervals.}
    \item{\code{parametric}}{Boolean indicating whether the bootstrap was parametric or non-parametric.}
    \item{\code{nr_simulations}}{The number of bootstrap samples used to create the simulated confidence envelopes.}
    \item{\code{iterations}}{The number of EM iterations, returned when the inflation and truncated negative binomial regression model cannot be separated.}
}

A itnb.ci-object is a list of class \code{itnb.ci} containing the following:
\describe{
    \item{\code{ci}}{A list of elements corresponding to the mean, overdispersion, and inflation proportion. If a confidence level is specified it contains the quantiles for each parameter, otherwise all estimated parameters are returned.}
    \item{\code{parametric}}{Boolean indicating whether the bootstrap was parametric or non-parametric.}
    \item{\code{level}}{The specified confidence level; if a confidence level is not supplied it is \code{NA}.}
    \item{\code{nr_simulations}}{The number of bootstrap samples used to create the simulated confidence envelopes.}
}
}
