% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb.R
\name{kitnb_optimisation}
\alias{kitnb_optimisation}
\title{kitnb EM optimisation}
\usage{
kitnb_optimisation(x, k, t, control = kitnb_optimisation.control())
}
\arguments{
\item{x}{observed values}

\item{k}{the point of inflation}

\item{t}{the point of truncation}

\item{control}{a control object, see \link{kitnb_optimisation.control} for details}
}
\value{
a list of the estimated parameters.
}
\description{
EM algorithm for kitnb parameter optimisation.
}
