% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktnb.R
\name{ktnb_optimisation}
\alias{ktnb_optimisation}
\title{Parameter optimisation.}
\usage{
ktnb_optimisation(x, k, t, control = ktnb_optimisation_control())
}
\arguments{
\item{x}{Observed data (numeric/integers).}

\item{k}{The point of inflation.}

\item{t}{The point of truncation.}

\item{control}{A control object, see \link{ktnb_optimisation_control} for details.}
}
\value{
A list of the estimated parameters, (mu, theta, p).
}
\description{
EM algorithm for ktnb parameter optimisation.
}
\examples{
library("ktnb")

n = 2000
k = 94
t = 93
mu = 100
theta = 10
p = 0.2

x <- rktnb(n, k, t, mu, theta, p)
ktnb_object <- ktnb_optimisation(x, k, t)
}
