% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktnb.R
\name{ktnb_optimisation_control}
\alias{ktnb_optimisation_control}
\title{Control function for the \link{ktnb_optimisation} function.}
\usage{
ktnb_optimisation_control(trace = TRUE, tolerance = 1e-06,
  iteration_max = 10000, iteration_min = 5, save_trace = FALSE)
}
\arguments{
\item{trace}{TRUE/FALSE show trace?}

\item{tolerance}{Convergence tolerance.}

\item{iteration_max}{The maximum number of allowed iterations.}

\item{iteration_min}{The minimum number of allowed iterations.}

\item{save_trace}{TRUE/FALSE should the entire trace be stored?}
}
\value{
A list of default arguments for the \link{ktnb_optimisation} function.
}
\description{
Creates a list of default options.
}
