% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-bootstrap.R
\name{lrtest}
\alias{lrtest}
\alias{lrtest.itnb}
\title{Likelihood ratio tests}
\usage{
lrtest(object, type = "overdispersion", level = 0.05, ...)

\method{lrtest}{itnb}(object, type = "overdispersion", level = 0.05, ...)
}
\arguments{
\item{object}{An \link{itnb-object}.}

\item{type}{String: The likelihood ratio test performed, either \code{"overdispersion"}, \code{"inflation"}, or \code{"both"}.}

\item{level}{Numeric: The significance level.}

\item{...}{Additional parameters (see details).}
}
\value{
An \link{lrtest-object}.
}
\description{
Likelihood ratio tests for the need of inflation and/or overdispersion in inflated and truncated negative binomial regression models.
}
\details{
...
}
