% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb.R
\name{pkitnb}
\alias{pkitnb}
\title{pkitnb}
\usage{
pkitnb(q, k, t = NULL, mu, theta, p, return.log = FALSE)
}
\arguments{
\item{q}{A vector of quantiles.}

\item{k}{The inflation point.}

\item{t}{The truncation point.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}
}
\value{
The probability mass.
}
\description{
The probability mass function of the k-inflated, t-t, negative binomial distribution.
}
