% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-bootstrap.R
\name{plot.ci}
\alias{plot.ci}
\title{Quantiles of bootstrapped \link{itnb-object}}
\usage{
\method{plot}{ci}(x, which = NULL, ...)
}
\arguments{
\item{x}{\link{ci-object}.}

\item{which}{String: Indicating the column of the trace to be shown (i.e.\ the log-likelihood or the name of a parameter). If left \code{NULL}, the function shows all traces.}

\item{...}{Additional arguments passed to the \link[graphics]{hist} function.}
}
\description{
A function plotting quantiles of parameter estimates returned from the \link{simulate_ci} function.
}
