% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-optimise.R
\name{residuals.itnb}
\alias{residuals.itnb}
\title{Residuals for inflated and truncated negative binomial regression models}
\usage{
\method{residuals}{itnb}(object, ...)
}
\arguments{
\item{object}{\link{itnb-object}.}

\item{...}{Additional arguments (see details).}
}
\description{
Residual function of an \link{itnb-object} (an object of class \code{itnb}).
}
\details{
The additional argument used by the function is '\code{type}'; taking a string, either \code{"pearson"}, \code{"deviance"}, \code{"response"}, \code{"link"}, specifying the scale
of the residuals.
}
