% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-rdp.R
\name{ritnb}
\alias{ritnb}
\title{ritnb}
\usage{
ritnb(n, mu, alpha, p = NULL, i = NULL, t = NULL)
}
\arguments{
\item{n}{Numeric (> 0): The number of observations.}

\item{mu}{Numeric (> 0): The expected value.}

\item{alpha}{Numeric (>= 0): The overdispersion.}

\item{p}{Numeric (>= 0): The inflation proportion.}

\item{i}{Numeric (>= 0): The inflation point.}

\item{t}{Numeric (>= 0): The truncation point.}
}
\value{
A vector of size \code{n} containing realisations of a itnb distribution.
}
\description{
Generates \code{n} random numbers from a i-inflated t-truncated negative binomial distribution with parameters \code{mu}, \code{alpha}, and \code{p} which is inflated at \code{i} and truncated at \code{t}.
}
