% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb.R
\name{rkitnb}
\alias{rkitnb}
\title{rkitnb}
\usage{
rkitnb(n, k, t = NULL, mu, theta, p)
}
\arguments{
\item{n}{The number of observations.}

\item{k}{The inflation point.}

\item{t}{The truncation point.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}
}
\value{
Random generated sample.
}
\description{
Generates random numbers from a k-inflated, t-t, negative binomial distribution.
}
