% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb.R
\name{rtnbinom}
\alias{rtnbinom}
\title{rtnbinom}
\usage{
rtnbinom(n, t, mu, theta)
}
\arguments{
\item{n}{The number of generated samples.}

\item{t}{The truncation point.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}
}
\value{
Random generated sample.
}
\description{
Generates numbers from a t-t negative binomial distribution
}
