% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-itnb.R
\name{simulate_ci}
\alias{simulate_ci}
\alias{simulate_ci.itnb}
\title{Confidence envelopes of itnb-object.}
\usage{
simulate_ci(
  object,
  level = 0.95,
  trace = TRUE,
  nr_simulations = 2000,
  parametric = FALSE,
  plot = FALSE
)

\method{simulate_ci}{itnb}(
  object,
  level = 0.95,
  trace = TRUE,
  nr_simulations = 2000,
  parametric = FALSE,
  plot = FALSE
)
}
\arguments{
\item{object}{An \link{itnb-object}.}

\item{level}{Numeric: The confidence level. If left as \code{NULL} all parametric bootstrap simulations are returned.}

\item{trace}{TRUE/FALSE: should a trace be shown?}

\item{nr_simulations}{Numeric: The number of simulations used to create the confidence envelopes.}

\item{parametric}{TRUE/FALSE: should the envelopes be simulated using the parametric bootstrap?}

\item{plot}{TRUE/FALSE: should a density plot of the parameters be returned?}
}
\value{
If \code{level = NULL} a matrix with bootstrap simulations, otherwise a matrix of lower and upper confidence limits for each parameter.
}
\description{
Simulated confidence envelopes of the parameters estimated by the \link{em_itnb} function.
}
\examples{
n = 2000
i = 94
t = 93
mu = 100
theta = 10
p = 0.2

x <- ritnb(n = n, mu = mu, theta = theta, p = p, i = i, t = t)
m <- em_itnb(x = x, i = i, t = t, control = em_itnb_control(trace = 100))
}
