% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-itnb.R
\name{simulate_confidence_envelopes}
\alias{simulate_confidence_envelopes}
\title{Confidence envelopes of itnb-object.}
\usage{
simulate_confidence_envelopes(
  itnb_object,
  level = 0.95,
  trace = TRUE,
  number_of_simulations = 100,
  plot_simulations = FALSE
)
}
\arguments{
\item{itnb_object}{An \link{itnb-object}.}

\item{level}{Numeric: The confidence level. If left as \code{NULL} all parametric bootstrap simulations are returned.}

\item{trace}{TRUE/FALSE, should a trace be shown?}

\item{number_of_simulations}{Integer: The number of simulations used to create the confidence envelopes.}

\item{plot_simulations}{TRUE/FALSE, should a density plot of the parameters be returned?}
}
\value{
A matrix with parametric bootstrap simulations (if \code{level = NULL}), or a matrix lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % (by default 2.5% and 97.5%).
}
\description{
Simulated confidence envelopes of the parameters estimated by the \link{em_itnb} function.
}
\examples{
n = 2000
i = 94
t = 93
mu = 100
theta = 10
p = 0.2

x <- ritnb(n = n, mu = mu, theta = theta, p = p, i = i, t = t)
m <- em_itnb(x = x, i = i, t = t, control = em_itnb_control(trace = 100))
}
