% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-optimise.R
\name{summary.itnb}
\alias{summary.itnb}
\title{Summary of an inflated and truncated negative binomial regression model fit}
\usage{
\method{summary}{itnb}(object, ...)
}
\arguments{
\item{object}{\link{itnb-object}.}

\item{...}{Additional arguments see details.}
}
\value{
A function summary.
}
\description{
Summary function for an \link{itnb-object} (an object of class \code{itnb}).
}
\details{
FIX FIX FIX
 are passed to the \link{itnb} and \link{confint.itnb}
}
