% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-bootstrap.R
\name{confint.itnb}
\alias{confint.itnb}
\title{Confidence intervals of itnb-object.}
\usage{
\method{confint}{itnb}(
  object,
  level = 0.95,
  nr_simulations = 200,
  parametric = FALSE,
  trace = 0,
  control = list()
)
}
\arguments{
\item{object}{An \link{itnb-object}.}

\item{level}{Numeric: The confidence level. If left as \code{NULL} all parametric bootstrap simulations are returned.}

\item{nr_simulations}{Numeric: The number of simulations used to create the confidence envelopes.}

\item{parametric}{TRUE/FALSE: should the envelopes be simulated using the parametric bootstrap?}

\item{trace}{Numeric (>= 0): showing a trace every \code{trace} number of iterations.}

\item{control}{List: A control object, see \link{itnb_control} for details, passed to the \link{itnb} function.}
}
\value{
If \code{level = NULL} a matrix with bootstrap simulations, otherwise a matrix of lower and upper confidence limits for each parameter.
}
\description{
Bootstrap confidence intervals of the parameters estimated by the \link{itnb} function.
}
\examples{
#
n <- 2000
i <- 2
t <- 1

#
x <- sort(runif(n, 0, 4))

beta <- c(3, 2)
mu <- exp(cbind(1, x) \%*\% beta)

theta <- 10
p <- 0.2

#
y <- ritnb(n = n, mu = mu, theta = theta, p = p, i = i, t = t)
data <- data.frame(y = y, x = x)

#
m <- itnb(
    y ~ x,
    data = data,
    i = i,
    t = t,
    link = "log"
)

#
confint(
    object = m,
    level = 0.95,
    nr_simulations = 20,
    parametric = FALSE
)
}
