% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{ditnb}
\alias{ditnb}
\title{ditnb}
\usage{
ditnb(x, mu, theta, p, i, t, lower_tail = TRUE, return_log = FALSE)
}
\arguments{
\item{x}{Numeric: A vector of quantiles.}

\item{mu}{Numeric: The expected value.}

\item{theta}{Numeric: The overdispersion.}

\item{p}{Numeric: The inflation proportion.}

\item{i}{Numeric: The inflation point.}

\item{t}{Numeric (< i): The truncation point (NB: can be set to \code{NULL}).}

\item{lower_tail}{TRUE/FALSE: should \eqn{P[X \leq x]} be returned in favour of\eqn{P[X \geq x]}?}

\item{return_log}{TRUE/FALSE: should the logarithm of the probabilities be returned?}
}
\value{
A vector the size as \code{x} containing the probability of each value.
}
\description{
The probability of each element in a vector \code{x} following the i-inflated t-truncated negative binomial distribution with parameters \code{mu}, \code{theta}, and \code{p} which is inflated at \code{i} and truncated at \code{t}.
}
