% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{ditnb}
\alias{ditnb}
\title{ditnb}
\usage{
ditnb(x, k, t = NULL, mu, theta, p, return_log = FALSE)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{k}{The inflation point.}

\item{t}{The truncation point.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}

\item{return_log}{TRUE/FALSE, should the log pmf be returned?}
}
\value{
A vector.
}
\description{
The density of a vector x following the k-inflated t-truncated negative binomial distribution.
}
