% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-itnb.R
\name{em_itnb_control}
\alias{em_itnb_control}
\title{Control function for the \link{em_itnb} function.}
\usage{
em_itnb_control(
  trace = 0,
  tolerance = 1e-06,
  iteration_min = 5,
  iteration_max = 10000,
  save_trace = FALSE
)
}
\arguments{
\item{trace}{Numeric (>= 0): showing a trace every \code{trace} number of iterations.}

\item{tolerance}{Numeric: Convergence tolerance.}

\item{iteration_min}{Numeric (>= 0): The minimum number of allowed iterations.}

\item{iteration_max}{Numeric (>= \code{iteration_min}): The maximum number of allowed iterations.}

\item{save_trace}{TRUE/FALSE: should the entire trace be stored?}
}
\value{
A list of default arguments for the \link{em_itnb} function.
}
\description{
Creates a list of default options.
}
