% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\name{itnb-object}
\alias{itnb-object}
\title{An itnb-object}
\description{
An itnb-object is a list containing the following:
\describe{
    \item{\code{formula}}{The \link{formula} supplied to the model.}
    \item{\code{data}}{If \code{save_trace = TRUE} in the control object, it contains the data used to optimise the parameters of the model, otherwise it is left \code{NA}.}
    \item{\code{i}}{The specified inflation point.}
    \item{\code{t}}{The specified truncation point.}
    \item{\code{loglikelihood}}{The loglikelihood of the optimised parameters.}
    \item{\code{beta}}{The optimised regression coefficients.}
    \item{\code{theta}}{The optimised overdispersion.}
    \item{\code{p}}{The optimised inflation proportion.}
    \item{\code{trace}}{If \code{save_trace = TRUE} in the control object, it contains a \link{data.frame} of the trace produced by the optimisation routine, otherwise it is set to \code{NA}.}
    \item{\code{converged}}{Boolean indicating whether the optimisation routine did or did not converge.}
    \item{\code{flag}}{If the optimisation routine did not converge, a string indicating the potential point of failure is returned, otherwise it is set to \code{NA}.}
}
}
