% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\docType{package}
\name{itnb-package}
\alias{itnb}
\alias{itnb-package}
\title{itnb: I-Inlfated T-Truncated Negative Binomial Distribution}
\description{
Very simple models for handling inflated and truncated over-dispersed count data.
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
\keyword{internal}
