% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\docType{package}
\name{itnb-package}
\alias{itnb}
\alias{itnb-package}
\title{itnb: Inlfated Truncated Negative Binomial Regression}
\description{
An inflated and truncated negative binomial regression model, with parameter estimation by the EM-algorithm, and confidence envelopes by either the parametric and non-parametric bootstrap.
}
\author{
\strong{Maintainer}: Søren B. Vilsen \email{svilsen@math.aau.dk}


Søren B. Vilsen <svilsen@math.aau.dk>
}
\keyword{internal}
