% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\name{itnb-package}
\alias{itnb-package}
\title{ITNB: Inflated and truncated negative binomial distribution}
\description{
ITNB: Inflated and truncated negative binomial distribution
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
