% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-package.R
\docType{package}
\name{itnb-package}
\alias{itnb}
\alias{itnb-package}
\title{itnb: I-Inlfated T-Truncated Negative Binomial Distribution}
\description{
An inflated and truncated negative binomial distribution parameterised using mean and overdispersion, with parameter estimation by the EM-algorithm, and confidence envelopes by both parametric and non-parametric bootstrap.
}
\author{
\strong{Maintainer}: Søren B. Vilsen \email{svilsen@math.aau.dk}


Søren B. Vilsen <svilsen@math.aau.dk>
}
\keyword{internal}
