% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-optimise.R
\name{itnb}
\alias{itnb}
\alias{itnb.formula}
\title{itnb parameter optimisation}
\usage{
itnb(formula, data = NULL, i = NULL, t = NULL, link = "log", control = list())

\method{itnb}{formula}(formula, data = NULL, i = NULL, t = NULL, link = "log", control = list())
}
\arguments{
\item{formula}{\link{formula}-object.}

\item{data}{\link{data.frame} (or \link[tibble]{tibble}): A data-set used to estimate the parameters.}

\item{i}{Numeric (i >= 0): The inflation point.}

\item{t}{Numeric (t >= 0): The truncation point.}

\item{link}{String: Indicating the link function, set to either 'identity', 'sqrt', or 'log' (default).}

\item{control}{List: A control object, see \link{itnb_control} for details.}
}
\value{
An object of class \link{itnb-object}.
}
\description{
Parameter optimisation for data generated by a r.v. following an itnb distribution using the ENM algorithm.
}
\examples{
#
n <- 2000
i <- 2
t <- 1

#
x <- sort(runif(n, 0, 4))

beta <- c(3, 2)
mu <- exp(cbind(1, x) \%*\% beta)

theta <- 10
p <- 0.2

#
y <- itnb::ritnb(n = n, mu = mu, theta = theta, p = p, i = i, t = t)
data <- data.frame(y = y, x = x)

#
itnb(
    y ~ x,
    data = data,
    i = i,
    t = t,
    link = "log"
)
}
