% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-optimise.R
\name{itnb}
\alias{itnb}
\alias{itnb.formula}
\title{itnb parameter optimisation}
\usage{
itnb(
  formula,
  data = NULL,
  i = NULL,
  t = NULL,
  link = "identity",
  control = list()
)

\method{itnb}{formula}(
  formula,
  data = NULL,
  i = NULL,
  t = NULL,
  link = "identity",
  control = list()
)
}
\arguments{
\item{formula}{A \link{formula}-object.}

\item{data}{A data-set, supplied as a \link{data.frame} (or \link[tibble]{tibble}).}

\item{i}{Numeric (i >= 0): The inflation point.}

\item{t}{Numeric (t >= 0): The truncation point.}

\item{link}{String indicating the link function, currently limited to 'log', 'sqrt', or 'identity'.}

\item{control}{List: A control object, see \link{itnb_control} for details.}
}
\value{
An object of class \link{itnb-object}.
}
\description{
Parameter optimisation for data generated by a r.v. following an itnb distribution using the ENM algorithm.
}
\examples{
n = 2000
i = 94
t = 93
mu = 100
theta = 10
p = 0.2

x <- ritnb(n = n, mu = mu, theta = theta, p = p, i = i, t = t)
em_itnb(x = x, i = i, t = t, control = em_itnb_control(trace = 100))
}
