% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{itnb_optimisation}
\alias{itnb_optimisation}
\title{Parameter optimisation.}
\usage{
itnb_optimisation(x, k, t, control = itnb_optimisation_control())
}
\arguments{
\item{x}{Observed data (numeric/integers).}

\item{k}{The point of inflation.}

\item{t}{The point of truncation.}

\item{control}{A control object, see \link{itnb_optimisation_control} for details.}
}
\value{
A list of the estimated parameters, (mu, theta, p).
}
\description{
EM algorithm for itnb parameter optimisation.
}
\examples{
library("ktnb")

n = 2000
k = 94
t = 93
mu = 100
theta = 10
p = 0.2

x <- rktnb(n, k, t, mu, theta, p)
ktnb_object <- ktnb_optimisation(x, k, t)
}
