% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{pitnb}
\alias{pitnb}
\title{pitnb}
\usage{
pitnb(q, mu, theta, p, i, t = NULL, return_log = FALSE)
}
\arguments{
\item{q}{A vector of quantiles.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}

\item{i}{The inflation point.}

\item{t}{The truncation point.}

\item{return_log}{TRUE/FALSE, should log of the cdf be returned?}
}
\value{
The probability mass.
}
\description{
The probability mass function of the i-inflated t-truncated negative binomial distribution.
}
