% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-rdp.R
\name{pitnb}
\alias{pitnb}
\title{pitnb}
\usage{
pitnb(
  q,
  mu,
  theta,
  p = NULL,
  i = NULL,
  t = NULL,
  lower_tail = TRUE,
  return_log = FALSE
)
}
\arguments{
\item{q}{Numeric: A vector of quantiles.}

\item{mu}{Numeric: The expected value.}

\item{theta}{Numeric: The overdispersion.}

\item{p}{Numeric: The inflation proportion.}

\item{i}{Numeric: The inflation point.}

\item{t}{Numeric: The truncation point.}

\item{lower_tail}{TRUE/FALSE: should \eqn{P[X \leq x]} be returned in favour of \eqn{P[X \geq x]}?}

\item{return_log}{TRUE/FALSE: should log of the cmf be returned?}
}
\value{
A vector the size as \code{q} containing the cumulative probability of each value.
}
\description{
The cumulative probability of each element in a vector \code{x} following the i-inflated t-truncated negative binomial distribution with parameters \code{mu}, \code{theta}, and \code{p} which is inflated at \code{i} and truncated at \code{t}.
}
