% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb-optimise.R
\name{plot.itnb}
\alias{plot.itnb}
\title{Plot parameter trace of an \link{itnb-object}}
\usage{
\method{plot}{itnb}(x, ...)
}
\arguments{
\item{x}{An \link{itnb-object}.}

\item{...}{Additional arguments passed to the \link[graphics]{plot} function.}
}
\description{
A function plotting the EM parameter trace returned by the \link{itnb} function. Note: the function can only be used if the argument \code{save_trace} in the \link{itnb_control} function was set to \code{TRUE}.
}
