% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{ritnb}
\alias{ritnb}
\title{ritnb}
\usage{
ritnb(n, k, t = NULL, mu, theta, p)
}
\arguments{
\item{n}{The number of observations.}

\item{k}{The inflation point.}

\item{t}{The truncation point.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}
}
\value{
Random generated sample.
}
\description{
Generates random numbers from a K-inflated t-truncated negative binomial distribution.
}
