% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{ritnb}
\alias{ritnb}
\title{ritnb}
\usage{
ritnb(n, mu, theta, p, i, t = NULL)
}
\arguments{
\item{n}{The number of observations.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{p}{The proportion of inflation.}

\item{i}{The inflation point.}

\item{t}{The truncation point.}
}
\value{
Random generated sample.
}
\description{
Generates random numbers from a i-inflated t-truncated negative binomial distribution.
}
