% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{ritnb}
\alias{ritnb}
\title{ritnb}
\usage{
ritnb(n, mu, theta, p, i, t)
}
\arguments{
\item{n}{Numeric: The number of observations.}

\item{mu}{Numeric: The expected value.}

\item{theta}{Numeric: The overdispersion.}

\item{p}{Numeric: The inflation proportion.}

\item{i}{Numeric: The inflation point.}

\item{t}{Numeric: The truncation point.}
}
\value{
A vector of size \code{n} containing realisations of a itnb distribution.
}
\description{
Generates \code{n} random numbers from a i-inflated t-truncated negative binomial distribution with parameters \code{mu}, \code{theta}, and \code{p} which is inflated at \code{i} and truncated at \code{t}.
}
