% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itnb.R
\name{rtnbinom}
\alias{rtnbinom}
\title{rtnbinom}
\usage{
rtnbinom(n, mu, theta, t)
}
\arguments{
\item{n}{The number of generated samples.}

\item{mu}{The expected value.}

\item{theta}{The overdispersion.}

\item{t}{The truncation point.}
}
\value{
Random generated sample.
}
\description{
Generates numbers from a t-truncated negative binomial distribution.
}
